Fix warning about converting from uint/int.

+++ src/core_write.cpp
@@ -30,7 +30,7 @@ UniValue ValueFromAmount(const CAmount& amount)
 
 UniValue ValueFromCapacity(const uint64_t& capacityTB)
 {
-    if (capacityTB < 10 * 1024) {
+    if (capacityTB < static_cast<uint64_t>(10 * 1024)) {
         return std::to_string(capacityTB) + " TB";
     } else {
         return std::to_string(capacityTB / 1024) + " PB";
